pyproject.toml
[tool.poetry]
name = "poetry-todo-api"
version = "0.1.0"
description = ""
authors = ["MoosaAfzal2 <moosaafzal2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
sqlmodel = "^0.0.16"
python-multipart = "^0.0.9"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^2.6.4"}
python-jose = "^3.3.0"
psycopg = {extras = ["binary"], version = "^3.1.18"}
bcrypt = "^4.1.2"
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "0.23.6"
mypy = "^1.9.0"
types-python-jose = "^3.3.4.20240106"
types-passlib = "^1.7.7.20240106"

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning::",  # ignore deprecations from all modules
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

